requirements.txt
python-telegram-bot[webhooks]==20.3
flask[async]==2.3.3
python-dotenv==1.0.1
pillow
pytesseract
openai==0.28.0
